requirements.txt
fastapi
uvicorn
httpx
jinja2
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
openai